moved delete functions to there corresponding components
